pyproject.toml
[project]
name = "xyzbank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "allure-pytest>=2.15.0",
    "faker>=37.12.0",
    "pytest>=8.4.2",
    "pytest-xdist>=3.8.0",
    "selenium>=4.38.0",
    "webdriver-manager>=4.0.2",
]
